Added logic for calculating ingredient amounts

+++ src/Ingredient.java
@@ -1,6 +1,6 @@
 public class Ingredient { // The attributes used to identify each ingredient
     public String name; // The name of the ingredient (Can include description ie. Minced, Stale, Toasted, etc.)
-    public Int Quantity; // The quantity of units that are required of this ingredient
+    public double Quantity; // The quantity of units that are required of this ingredient
     public Units Unit; // The unit that the quantity refers to.  View Units.java for reference.
 
     public Ingredient ConvertUnits(Units unitToConvertTo) {
